consumer/pyproject.toml
[tool.poetry]
name = "consumer"
version = "0.1.0"
description = ""
authors = ["specnazm <laketic.milena98@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
faust = "^1.10.4"
stockfish = "^3.23.0"
cassandra-driver = "^3.25.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

